4 pin mode working!
The issue with 4-pin mode is that it is expecting synchronous inputs, and the Teensy is asynchronous.  One implementation got all but '0' and '1' to render correctly, while the for loop implementation was missing many other characters but had '0' and '1'.  Therefore, by using the for loop method for '0' and '1', and the literal bit-mask method for all other characters, the display behaves correctly.